+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+### 0.4.0 (2017-12-16)
+
+- Removing multiple requests, now has a queue with 1s delay
+- Adding 5 minutes cache to every request
+- Adding partial absenses scrapper
+- Solving #4 by adding partial enrolled disciplines method scrapper
+
 ### 0.3.0 (2017-12-15)
 
-Solved #2 by adding partial grades scrapper
+- Solved #2 by adding partial grades scrapper
 
 ### 0.2.0 (2017-12-15)
 
-Solved #1 by adding registered emails scrapper
+- Solved #1 by adding registered emails scrapper